--- app/src/main/java/ru/iteco/fmhandroid/dao/ClaimCommentDao.kt
@@ -1,21 +1,0 @@
-package ru.iteco.fmhandroid.dao
-
-import androidx.room.Dao
-import androidx.room.Insert
-import androidx.room.OnConflictStrategy
-import androidx.room.Query
-import kotlinx.coroutines.flow.Flow
-import ru.iteco.fmhandroid.dto.ClaimComment
-import ru.iteco.fmhandroid.entity.ClaimCommentEntity
-
-@Dao
-interface ClaimCommentDao {
-    @Query("SELECT * FROM ClaimCommentEntity WHERE claimId = :claimId")
-    fun getClaimComments(claimId: Int): Flow<List<ClaimComment>>
-
-    @Insert(onConflict = OnConflictStrategy.REPLACE)
-    suspend fun insertComments(comments: List<ClaimCommentEntity>)
-
-    @Insert(onConflict = OnConflictStrategy.REPLACE)
-    suspend fun insertComment(comment: ClaimCommentEntity)
-}
