+++ data/db/src/main/kotlin/ru/iteco/fmh/data/db/entity/NewsCategoryEntity.kt
@@ -1,0 +1,16 @@
+package ru.iteco.fmh.data.db.entity
+
+import androidx.room.ColumnInfo
+import androidx.room.Entity
+import androidx.room.PrimaryKey
+
+@Entity(tableName = "NewsCategoryEntity")
+data class NewsCategoryEntity(
+    @PrimaryKey
+    @ColumnInfo(name = "id")
+    val id: Int,
+    @ColumnInfo(name = "name")
+    val name: String,
+    @ColumnInfo(name = "deleted")
+    val deleted: Boolean
+)
